4.arm_raspberry/requirements.txt
numpy==2.2.4
opencv-python==4.11.0.86
RPi.GPIO==0.7.1